requirements.txt
torch>=1.1.0
torchvision>=0.3.0
visdom>=0.1.8.9
PyYAML>=6.0
scikit-learn>=0.23.2
dominate>=2.5.1
opencv-python>=4.4.0.42
matplotlib>=3.3.1
Pillow>=7.2.0